#8 Unable to decrypt zip file that was created with StandardEncryption

+++ src/main/java/net/lingala/zip4j/tasks/AbstractAddFileToZipTask.java
@@ -168,7 +168,7 @@ public abstract class AbstractAddFileToZipTask<T> extends AsyncZipTask<T> {
     } else {
       if (clonedZipParameters.isEncryptFiles() && clonedZipParameters.getEncryptionMethod() == ZIP_STANDARD) {
         progressMonitor.setCurrentTask(CALCULATE_CRC);
-        clonedZipParameters.setEntryCRC((int) computeFileCrc(fileToAdd, progressMonitor));
+        clonedZipParameters.setEntryCRC(computeFileCrc(fileToAdd, progressMonitor));
         progressMonitor.setCurrentTask(ADD_ENTRY);
       }
